+++ statsmodels/sandbox/unfinished/example_svar.py
