+++ statsmodels/sandbox/unfinished/smoothers_lowess_old.py
