+++ statsmodels/sandbox/unfinished/test_summary.py
