+++ statsmodels/sandbox/unfinished/test_summary_old.py
